+++ base/src/main/scala/org/opentorah/xml/Catalog.scala
@@ -7,11 +7,11 @@ object Catalog extends Dialect, Doctype:
 
   override val mimeType: String = "application/xml"
 
-  val dtdId: String = "-//OASIS//DTD XML Catalogs V1.1//EN"
-
-  val dtdUri: String = "http://www.oasis-open.org/committees/entity/release/1.1/catalog.dtd"
-
-  override val doctype: String = s"""<!DOCTYPE catalog PUBLIC "$dtdId" "$dtdUri">"""
+  override val doctype: String = Doctype.string(
+    "catalog",
+    dtdId = "-//OASIS//DTD XML Catalogs V1.1//EN",
+    dtdUri = "http://www.oasis-open.org/committees/entity/release/1.1/catalog.dtd"
+  )
 
   def write(
     file: File,